backend/requirements.txt
# FastAPI and Web
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
httpx>=0.25.1
python-jose>=3.3.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.23
asyncpg>=0.28.0
alembic>=1.12.1

# Utilities
python-dotenv>=1.0.0
asyncio>=3.4.3
passlib>=1.7.4
bcrypt>=4.0.1

# Date/Time handling
pytz>=2023.3
schedule>=1.2.1
croniter>=1.4.1

# Logging and Monitoring
python-json-logger>=2.0.7
PyYAML>=6.0.1
psutil>=5.9.6
prometheus-client>=0.17.1
prometheus-fastapi-instrumentator>=6.1.0
statsd>=4.0.1
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
sentry-sdk>=1.38.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
